refactor (front): separando a logica do authService em funcoes
